# Launch tests in CTest with exact amount of ranks