## Main changes of this PR

This PR adds the amount of requested ranks per test to the output of `./testprecice --list_units` and launches the tests either without MPI (if only 1 rank is requested) or with the exact amount of required MPI ranks.

## Motivation and additional information

* This allows using any amount of ranks in tests, removing the historic maximum of 4.
* Running tests via CTest are always correctly sized and don't need to handle unused ranks, resulting in more consistent runtimes.
* Tests using one rank should be able to run for builds with MPI disabled.

Less communication and initialization leads to a more consistent test runtime.
On my local machine:

```
Benchmark develop: ctest -j
  Time (mean ± σ):     28.418 s ±  4.859 s    [User: 317.758 s, System: 213.569 s]
  Range (min … max):   25.990 s … 38.546 s    10 runs

Benchmark this PR: ctest -j
  Time (mean ± σ):     20.106 s ±  0.828 s    [User: 208.488 s, System: 122.066 s]
  Range (min … max):   18.841 s … 21.008 s    10 runs

Summary
  this PR
    1.41 ± 0.25 times faster than develop
```

Further possible improvements:
* Don't initialize MPI running when using only 1 rank. This can lead to many problems though.
* ~~Use the amount of ranks as CTest processors to prevent overloading the system. I'll look into this.~~ Too conservative. 50% slower.

## Author's checklist

* [x] I used the [`pre-commit` hook](https://precice.org/dev-docs-dev-tooling.html#setting-up-pre-commit) to prevent dirty commits and used `pre-commit run --all` to format old commits.
* [ ] I added a changelog file with `make changelog` if there are user-observable changes since the last release.
* [ ] I added a test to cover the proposed changes in our test suite.
* [ ] For breaking changes: I documented the changes in the appropriate [porting guide](https://precice.org/couple-your-code-porting-overview.html).
* [x] I stuck to C++17 features.
* [x] I stuck to CMake version 3.22.1.
* [ ] I squashed / am about to squash all commits that should be seen as one.