@fsimonis 2024-11-29
> I must admit I don't understand how these changes give this effect and currently it is not tested as we don't have such a test yet, right?

Until now, CTest had to guess on how to many ranks to use to launch individual test suites. We picked 4 as it covered all test setups.

Now, CTest launches each test separately, meaning it doesn't need to satisfy the max ranks of an entire test suite.
Furthermore, CTest knows how many ranks are required by a test and launches it with the necessary amount. Hence, a test can request 6 ranks and CTest will launch it with 6.